--- app/model/SectionInfo.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class SectionInfo:
-    index: int
