+++ app/processing/graph.py
@@ -4,7 +4,7 @@ from openqasm3.ast import Program
 
 
 @dataclass(frozen=True)
-class ModelNode:
+class ProgramNode:
     """
     Represents a node in a visual model of an openqasm3 program.
     """
@@ -12,3 +12,9 @@ class ModelNode:
     id: str
     Implementation: Program
     UncomputeImplementation: Program | None = None
+
+
+@dataclass
+class SectionInfo:
+    index: int
+    node: ProgramNode
