+++ docs/processing/pre.rst
@@ -4,25 +4,25 @@ Preprocessing
 Each qasm snippet attached to a node in the editor will first be passed through the preprocessing pipeline.
 The pipeline consists of multiple :class:`openqasm3.visitor.QASMTransformer` that will transform the abstract syntax tree (AST) of the qasm snippet.
 
-.. automodule:: app.preprocessing
+.. automodule:: app.processing
    :members:
 
 In order to prevent collision while merging all the standalone qasm snippets, the preprocessing pipeline needs to know the position of the snippet in the final program.
 
-.. automodule:: app.model.SectionInfo
+.. automodule:: app.processing.graph.SectionInfo
    :members:
 
 Utils
 -----
 
-.. automodule:: app.preprocessing.utils
+.. automodule:: app.processing.utils
     :members:
     :undoc-members:
 
 Renaming
 --------
 
-.. automodule:: app.preprocessing.renaming
+.. automodule:: app.processing.pre.renaming
     :members:
     :undoc-members:
